API change: log and fft

+++ ZSTACKLIB/Algo/wavegen.cpp
@@ -4,10 +4,10 @@
 #define PI 3.1415927
 #define DOUBLE_PI   6.283185307179 //6.283185307179586476925286766559
 
-int wavegen(double *output, int length, int freq, int sample_freq)
+int wavegen(float *output, int length, int freq, int sample_freq)
 {
 	int i;
-	double value;
+	float value;
 
 	if (NULL == output)
 		return -1;
